requirements.txt
fastapi
uvicorn[standard]
transformers
jinja2
openai-whisper
pydub
ffmpeg-python
coqui-tts
numpy
httpx


